all features realised

+++ app/src/main/java/com/evilstan/starwarsuniverse/BaseViewModel.kt
@@ -1,51 +1,36 @@
 package com.evilstan.starwarsuniverse
 
-import androidx.lifecycle.MutableLiveData
+import android.content.Context
+import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.evilstan.starwarsuniverse.data.core.NetModule
-import com.evilstan.starwarsuniverse.data.core.ResponseWrapper
-import com.evilstan.starwarsuniverse.data.core.StarWarsApi
-import com.evilstan.starwarsuniverse.domain.Repository
 import com.evilstan.starwarsuniverse.domain.cache.PersonCache
-import kotlinx.coroutines.Dispatchers
+import com.evilstan.starwarsuniverse.domain.database.AppDatabase
+import com.evilstan.starwarsuniverse.domain.database.Repository
 import kotlinx.coroutines.launch
 
-abstract class BaseViewModel : ViewModel() {
+abstract class BaseViewModel(context: Context) : ViewModel() {
+    private val movieDao = AppDatabase.getInstance(context).personDao()
+    private val repository = Repository(movieDao)
+    val cachedPersons: LiveData<List<PersonCache>> = repository.getAllByLiveData()
 
-    var api: StarWarsApi = NetModule().service(StarWarsApi::class.java)
+    private var contains = false
 
-    fun <T> requestWithLiveData(
-        liveData: MutableLiveData<Event<T>>,
-        request: suspend () -> ResponseWrapper<T>
-    ) {
-        liveData.postValue(Event.loading())
+    suspend fun dbContains(name: String): Boolean {
+        viewModelScope.launch {
+            contains = repository.contains(name)
+        }.join()
+        return contains
+    }
 
-        this.viewModelScope.launch(Dispatchers.IO) {
-            try {
-                val response = request.invoke()
-                if (response.data != null) {
-                    liveData.postValue(Event.success(response.data))
-                }
-            } catch (e: Exception) {
-                e.printStackTrace()
-                liveData.postValue(Event.error(null))
+    fun makeFavorite(personCache: PersonCache, favorite: Boolean) {
+        personCache.favorite = favorite
+        viewModelScope.launch {
+            if (favorite) {
+                repository.insert(personCache)
+            } else {
+                repository.delete(personCache)
             }
         }
-    }
-
-    fun getPersons(repository: Repository, liveData: MutableLiveData<List<PersonCache>>) =
-        viewModelScope.launch {
-            val record = repository.allPersons
-            liveData.postValue(record.value)
-        }
-
-    fun insert(personCache: PersonCache, repository: Repository) = viewModelScope.launch {
-        repository.insert(personCache)
-        println("Records in db: " + repository.getCount())
-    }
-
-    fun delete(personCache: PersonCache, repository: Repository) = viewModelScope.launch {
-        repository.delete(personCache)
     }
 }
